aula_6/memory-ai/pyproject.toml
[project]
name = "memory-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Ian Vieira Salgado",email = "vieira4ev@gmail.com"}
]
readme = "README.md"
requires-python = "~3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-core (>=0.3.75,<0.4.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "langgraph-cli[inmem] (>=0.4.0,<0.5.0)"
]

[tool.poetry]
packages = [{include = "memory_ai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
